core_task/contract/serializers.py
from rest_framework import serializers

from contract.models import Contract


class ContractSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contract
        fields = "__all__"
